refactor(recipes): consolidate finetune utilities
