+++ skills/nemo-mbridge-perf-sequence-packing/card.yaml
@@ -80,7 +80,7 @@ evidence:
   - src/megatron/bridge/data/sequence_batching.py
   - src/megatron/bridge/training/vlm_step.py
   - src/megatron/bridge/training/config.py
-  - src/megatron/bridge/recipes/utils/finetune_utils.py
+  - src/megatron/bridge/recipes/utils/dataset_utils.py
   - src/megatron/bridge/recipes/common.py
   - src/megatron/bridge/recipes/llama/h100/llama3.py
   - src/megatron/bridge/recipes/qwen/h100/qwen3_next.py
