--- src/megatron/bridge/recipes/utils/finetune_utils.py
@@ -1,231 +1,0 @@
-# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Utility functions for finetuning recipes."""
-
-from typing import Any
-
-from megatron.bridge.data.builders import (
-    ChatSFTPreprocessingConfig,
-    GPTSFTDatasetConfig,
-    HFDatasetSourceConfig,
-    PromptCompletionSFTPreprocessingConfig,
-    SFTPreprocessingConfig,
-)
-from megatron.bridge.data.packing import PackedSequenceSpecs
-from megatron.bridge.peft.base import PEFT
-from megatron.bridge.peft.dora import DoRA
-from megatron.bridge.peft.lora import LoRA
-
-
-def default_peft_config(peft_scheme: str | PEFT | None, **kwargs) -> PEFT | None:
-    """Create default PEFT configuration matching NeMo2 exactly.
-
-    Args:
-        peft_scheme: PEFT scheme - 'lora', 'dora', PEFT instance, or None for full finetuning
-
-    Returns:
-        PEFT configuration or None for full finetuning
-    """
-    if peft_scheme is None:
-        return None  # Full finetuning
-
-    if isinstance(peft_scheme, PEFT):
-        return peft_scheme  # User provided custom PEFT
-
-    if isinstance(peft_scheme, str):
-        if peft_scheme.lower() == "none":
-            return None
-        if peft_scheme.lower() == "lora":
-            return LoRA(**kwargs)
-        elif peft_scheme.lower() == "dora":
-            return DoRA(**kwargs)
-        else:
-            raise ValueError(f"Unknown PEFT scheme: {peft_scheme}. Supported: 'lora', 'dora', or None")
-
-    raise ValueError(f"Invalid peft type: {type(peft_scheme)}. Expected str, PEFT instance, or None")
-
-
-def _text_hf_dataset_config(
-    *,
-    seq_length: int,
-    source: HFDatasetSourceConfig,
-    preprocessing: SFTPreprocessingConfig,
-    validation_source: HFDatasetSourceConfig | None = None,
-    test_source: HFDatasetSourceConfig | None = None,
-    do_validation: bool = True,
-    do_test: bool = False,
-    enable_offline_packing: bool = False,
-    offline_packing_specs: PackedSequenceSpecs | None = None,
-    dataset_kwargs: dict[str, Any] | None = None,
-    val_proportion: float | None = None,
-    num_workers: int = 2,
-) -> GPTSFTDatasetConfig:
-    """Create an HF-backed text SFT config with optional offline packing."""
-    return GPTSFTDatasetConfig(
-        seq_length=seq_length,
-        hf_dataset=source,
-        hf_validation_dataset=validation_source,
-        hf_test_dataset=test_source,
-        hf_validation_proportion=val_proportion,
-        do_validation=do_validation,
-        do_test=do_test,
-        preprocessing=preprocessing,
-        enable_offline_packing=enable_offline_packing,
-        offline_packing_specs=offline_packing_specs,
-        dataset_kwargs=dataset_kwargs,
-        seed=5678,
-        dataloader_type="batch",
-        num_workers=num_workers,
-        data_sharding=True,
-        pin_memory=True,
-        persistent_workers=False,
-    )
-
-
-def default_squad_config(
-    seq_length: int, enable_offline_packing: bool = True, pad_seq_to_mult: int = 1
-) -> GPTSFTDatasetConfig:
-    """Create default SQuAD dataset configuration for finetuning recipes.
-
-    Args:
-        seq_length: Sequence length for the dataset
-        enable_offline_packing: Whether to enable offline packed-sequence preparation.
-        pad_seq_to_mult: Optional multiple to pad each sequence to when packing
-            (set to `2 * context_parallel_size` for THD CP runs).
-
-    Returns:
-        GPTSFTDatasetConfig configured for SQuAD finetuning
-
-    Note:
-        Uses consistent settings across all finetuning recipes:
-        - SQuAD dataset with appropriate dataloader type
-        - 10% validation slice
-        - Seed 5678 (different from pretrain seed 1234)
-    """
-    dataset_kwargs = {}
-    offline_packing_specs = None
-    if enable_offline_packing:
-        dataset_kwargs["pad_to_max_length"] = True
-        offline_packing_specs = PackedSequenceSpecs(packed_sequence_size=seq_length, pad_seq_to_mult=pad_seq_to_mult)
-
-    return _text_hf_dataset_config(
-        source=HFDatasetSourceConfig(dataset_name="squad"),
-        preprocessing=PromptCompletionSFTPreprocessingConfig(separator=" "),
-        seq_length=seq_length,
-        enable_offline_packing=enable_offline_packing,
-        offline_packing_specs=offline_packing_specs,
-        dataset_kwargs=dataset_kwargs,
-        val_proportion=0.1,
-        num_workers=1,
-    )
-
-
-def default_openmathinstruct2_config(
-    seq_length: int = 4096,
-    enable_offline_packing: bool = False,
-    pad_seq_to_mult: int = 1,
-) -> GPTSFTDatasetConfig:
-    """Create the default OpenMathInstruct-2 finetuning dataset.
-
-    Args:
-        seq_length: Maximum sequence length.
-        enable_offline_packing: Whether to enable offline text SFT packing. This does
-            not enable VLM in-batch packing.
-        pad_seq_to_mult: Sequence-length multiple used by offline packing.
-
-    Returns:
-        OpenMathInstruct-2 dataset configuration.
-    """
-    offline_packing_specs = None
-    if enable_offline_packing:
-        offline_packing_specs = PackedSequenceSpecs(packed_sequence_size=seq_length, pad_seq_to_mult=pad_seq_to_mult)
-
-    return _text_hf_dataset_config(
-        source=HFDatasetSourceConfig(dataset_name="openmathinstruct2"),
-        preprocessing=PromptCompletionSFTPreprocessingConfig(separator=" "),
-        seq_length=seq_length,
-        enable_offline_packing=enable_offline_packing,
-        offline_packing_specs=offline_packing_specs,
-        val_proportion=0.05,
-        num_workers=2,
-    )
-
-
-def default_gsm8k_config(
-    seq_length: int = 2048,
-    enable_offline_packing: bool = False,
-    pad_seq_to_mult: int = 1,
-) -> GPTSFTDatasetConfig:
-    """Create default GSM8K dataset configuration for finetuning recipes.
-
-    GSM8K (Grade School Math 8K) is a dataset of 8.5K high quality linguistically diverse
-    grade school math word problems. See: https://huggingface.co/datasets/openai/gsm8k
-
-    Args:
-        seq_length: Sequence length for the dataset (default 2048, sufficient for GSM8K)
-        enable_offline_packing: Whether to enable offline text SFT packing. This does
-            not enable VLM in-batch packing.
-        pad_seq_to_mult: Optional multiple to pad each sequence to when packing.
-
-    Returns:
-        GPTSFTDatasetConfig configured for GSM8K finetuning
-
-    Note:
-        - GSM8K has 7,473 train and 1,319 test examples
-        - Loads the full DatasetDict so the published test split is used for evaluation
-    """
-    offline_packing_specs = None
-    if enable_offline_packing:
-        offline_packing_specs = PackedSequenceSpecs(packed_sequence_size=seq_length, pad_seq_to_mult=pad_seq_to_mult)
-
-    return _text_hf_dataset_config(
-        source=HFDatasetSourceConfig(dataset_name="gsm8k"),
-        preprocessing=PromptCompletionSFTPreprocessingConfig(separator=" "),
-        test_source=HFDatasetSourceConfig(dataset_name="gsm8k", split="test"),
-        do_validation=False,
-        do_test=True,
-        seq_length=seq_length,
-        enable_offline_packing=enable_offline_packing,
-        offline_packing_specs=offline_packing_specs,
-        num_workers=2,
-    )
-
-
-def default_openmathinstruct2_thinking_config(
-    seq_length: int = 4096,
-    enable_offline_packing: bool = False,
-    pad_seq_to_mult: int = 1,
-) -> GPTSFTDatasetConfig:
-    """Create OpenMathInstruct-2 dataset config with CoT in analysis channel, answer in final channel.
-
-    Puts generated_solution (minus the trailing \boxed{N}) into the assistant thinking field
-    (rendered as <|channel|>analysis) and #### {expected_answer} into the content field
-    (rendered as <|channel|>final).
-
-    Args:
-        seq_length: Sequence length (default 4096)
-        enable_offline_packing: Whether to enable offline text SFT packing. This does
-            not enable VLM in-batch packing.
-        pad_seq_to_mult: Padding multiple for packing (set to 2*CP for THD CP runs).
-    """
-    cfg = default_openmathinstruct2_config(
-        seq_length=seq_length,
-        enable_offline_packing=enable_offline_packing,
-        pad_seq_to_mult=pad_seq_to_mult,
-    )
-    assert cfg.hf_dataset is not None
-    cfg.hf_dataset = HFDatasetSourceConfig(dataset_name="openmathinstruct2_thinking")
-    cfg.preprocessing = ChatSFTPreprocessingConfig()
-    return cfg
